bug: fix debug format of IP in AuthForbidden err.
The `Error::AuthForbidden` variant is user-facing and rendered as a JSON
error response. Prev. it was using the debug format specifier for the
`LowerName` of the subdomain in the update request that was rejected.
This included superfluous type information. Removing the debug specifier
presents just the domain name in the error.
diff --git a/src/error.rs b/src/error.rs
index 9e30b55..9d53edf 100644
--- a/src/error.rs
+++ b/src/error.rs
@@ -7,7 +7,7 @@ use trust_dns_server::proto::error::ProtoError;
 pub enum Error {
     #[error("not implemented")]
     NotImplemented,
-    #[error("IP {0} is not authorized to update {1:?}")]
+    #[error("IP {0} is not authorized to update {1}")]
     AuthForbidden(IpAddr, LowerName),
     #[error(transparent)]
     JsonExtractorRejection(#[from] JsonRejection),