nix: add NixOS module.
This commit adds the start of a Flake exported NixOS module for ACME
Crab. This module allows using Nix to configure ACME Crab, producing
a JSON config file and hardened systemd unit for the service.

This is an OK starting point, but there's work remaining to tighten up
the validation (many elements are stringly-types) and the systemd unit
hardening could be taken further.